\name{e}
\alias{e}

\title{Environment for data and variables.}

\description{
Environment that contains variables object \code{FarmVars} and, if downloaded, full data set \code{FarmData}.
}

\usage{
e
}

\details{
\code{e} contains data and variables, see \code{\link{FarmVars}} and \code{\link{FarmData}}.
}

\value{
\code{e} returns the environment for \code{FarmVars} and \code{FarmData}.
}

\author{
Carsten Croonenbroeck
}

\seealso{
\code{\link{AcquireData}} for downloading the full data set.
}

\examples{
e
## Gives the environment object.
}

\keyword{environment}
\keyword{Data}
