\name{wflo}
\docType{package}
\alias{wflo}
\alias{wflo-package}

\title{Data set and functions for wind farm layout optimization.}

\description{
This package makes two contributions to the Wind Farm Layout Optimization (WFLO) research branch: First, it provides a convenient and realistic data set of high resolution and accuracy encompassing wind speeds, wind directions, standard deviations of wind directions, and (adjusted) yields for the entire area of Germany. Second, it supplies a set of helper functions and a benchmark function for economically (profit) driven wind farm layout optimization. This enables researchers in the field of the np-hard problem of wflo to focus on their optimization methodology contribution and also provides a realistic benchmark setting for comparability among contributions. 
}

\author{
Carsten Croonenbroeck\cr
David Hennecke
}

\keyword{packages}
